requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
asgiref==3.5.2; python_version >= '3.7'
django==4.0.6
djangorestframework==3.13.1
gunicorn==20.1.0
pytz==2022.1
setuptools==63.2.0; python_version >= '3.7'
sqlparse==0.4.2; python_version >= '3.5'
tzdata==2022.1; sys_platform == 'win32'
